+++ modules/hanna-utils/src/assets.ts
@@ -407,7 +407,7 @@ export const getBlingUrl = (blingType: BlingType): string =>
 
 // ===========================================================================
 
-type Favicon =
+export type Favicon =
   | 'favicon.svg'
   | 'favicon.png'
   | 'favicon-large.png' // same as favicon-144
@@ -424,3 +424,22 @@ type Favicon =
  * @see https://www.npmjs.com/package/@reykjavik/hanna-utils#favicons
  */
 export const getFavicon = (faviconFile: Favicon): string => getAssetUrl(faviconFile);
+
+export type RvkLogo =
+  | 'reykjavik-logo.svg'
+  | 'reykjavik-logo-notext.svg'
+  | 'reykjavik-logo.png'
+  | 'reykjavik-logo--large.png'
+  | 'reykjavik-logo--small.png'
+  | 'reykjavik-logo-notext.png'
+  | 'reykjavik-logo-notext--large.png'
+  | 'reykjavik-logo-notext--small.png';
+
+/**
+  Generates a URL to an image of Reykjavík's logo (i.e. the coat of arms)
+  on the style server.
+
+  @see https://www.npmjs.com/package/@reykjavik/hanna-utils#reykjavik-logo
+ */
+export const getRvkLogoUrl = (logoFile: RvkLogo = 'reykjavik-logo.svg'): string =>
+  getAssetUrl(logoFile);